[CAMEL-11807] Upgrade camel-corda to junit5
