+++ components/camel-corda/src/test/java/org/apache/camel/component/corda/CordaConsumerStartTrackedFlowDynamicIntegrationTest.java
@@ -18,20 +18,22 @@ package org.apache.camel.component.corda;
 
 import org.apache.camel.builder.RouteBuilder;
 import org.apache.camel.component.mock.MockEndpoint;
-import org.junit.Ignore;
-import org.junit.Test;
+import org.junit.jupiter.api.Test;
 
 import static org.apache.camel.component.corda.CordaConstants.OPERATION;
-import static org.apache.camel.component.corda.CordaConstants.STATE_MACHINE_FEED;
+import static org.apache.camel.component.corda.CordaConstants.START_TRACKED_FLOW_DYNAMIC;
+import static org.junit.jupiter.api.Assertions.assertEquals;
 
-@Ignore("This integration test requires a locally running corda node such cordapp-template-java")
-public class CordaConsumerStateMachineFeedTest extends CordaConsumerTestSupport {
+public class CordaConsumerStartTrackedFlowDynamicIntegrationTest extends CordaConsumerTestSupport {
+
 
     @Test
-    public void stateMachineFeedTest() throws Exception {
+    public void startTrackedFlowDynamicTest() throws Exception {
+        //Expects CamelFlow is deployed on the node
         mockResult.expectedMinimumMessageCount(1);
         mockError.expectedMessageCount(0);
         MockEndpoint.assertIsSatisfied(context);
+        assertEquals("Hello world!", mockResult.getExchanges().get(0).getIn().getBody());
     }
 
     @Override
@@ -40,7 +42,9 @@ public class CordaConsumerStateMachineFeedTest extends CordaConsumerTestSupport 
             public void configure() {
                 errorHandler(deadLetterChannel("mock:error"));
 
-                from(getUrl() + "&" + OPERATION.toLowerCase() + "=" + STATE_MACHINE_FEED)
+                from(getUrl() + "&" + OPERATION.toLowerCase() + "=" + START_TRACKED_FLOW_DYNAMIC
+                    + "&flowLociClass=#flowLociClass"
+                    + "&arguments=#arguments")
                         .to("mock:result");
             }
         };
