+++ components/camel-corda/src/test/java/org/apache/camel/component/corda/CordaConsumerTransactionMapFeedIntegrationTest.java
@@ -18,17 +18,15 @@ package org.apache.camel.component.corda;
 
 import org.apache.camel.builder.RouteBuilder;
 import org.apache.camel.component.mock.MockEndpoint;
-import org.junit.Ignore;
-import org.junit.Test;
+import org.junit.jupiter.api.Test;
 
 import static org.apache.camel.component.corda.CordaConstants.OPERATION;
-import static org.apache.camel.component.corda.CordaConstants.VAULT_TRACK_BY;
+import static org.apache.camel.component.corda.CordaConstants.STATE_MACHINE_RECORDED_TRANSACTION_MAPPING_FEED;
 
-@Ignore("This integration test requires a locally running corda node such cordapp-template-java")
-public class CordaConsumerVaultTrackByTest extends CordaConsumerTestSupport {
+public class CordaConsumerTransactionMapFeedIntegrationTest extends CordaConsumerTestSupport {
 
     @Test
-    public void vaultTrackByTest() throws Exception {
+    public void transactionMapFeedTest() throws Exception {
         mockResult.expectedMinimumMessageCount(1);
         mockError.expectedMessageCount(0);
         MockEndpoint.assertIsSatisfied(context);
@@ -40,11 +38,7 @@ public class CordaConsumerVaultTrackByTest extends CordaConsumerTestSupport {
             public void configure() {
                 errorHandler(deadLetterChannel("mock:error"));
 
-                from(getUrl() + "&" + OPERATION.toLowerCase() + "=" + VAULT_TRACK_BY
-                        + "&contractStateClass=#contractStateClass"
-                        + "&queryCriteria=#queryCriteria"
-                        + "&pageSpecification=#pageSpecification"
-                        + "&sort=#sort")
+                from(getUrl() + "&" + OPERATION.toLowerCase() + "=" + STATE_MACHINE_RECORDED_TRANSACTION_MAPPING_FEED)
                         .to("mock:result");
             }
         };
