+++ components/camel-corda/src/test/java/org/apache/camel/component/corda/CordaConsumerVaultTrackByCriteriaIntegrationTest.java
@@ -18,17 +18,15 @@ package org.apache.camel.component.corda;
 
 import org.apache.camel.builder.RouteBuilder;
 import org.apache.camel.component.mock.MockEndpoint;
-import org.junit.Ignore;
-import org.junit.Test;
+import org.junit.jupiter.api.Test;
 
 import static org.apache.camel.component.corda.CordaConstants.OPERATION;
-import static org.apache.camel.component.corda.CordaConstants.VAULT_TRACK_BY_WITH_SORTING;
+import static org.apache.camel.component.corda.CordaConstants.VAULT_TRACK_BY_CRITERIA;
 
-@Ignore("This integration test requires a locally running corda node such cordapp-template-java")
-public class CordaConsumerVaultTrackByWithSortingTest extends CordaConsumerTestSupport {
+public class CordaConsumerVaultTrackByCriteriaIntegrationTest extends CordaConsumerTestSupport {
 
     @Test
-    public void vaultTrackByWithSortingTest() throws Exception {
+    public void vaultTrackByCriteriaTest() throws Exception {
         mockResult.expectedMinimumMessageCount(1);
         mockError.expectedMessageCount(0);
         MockEndpoint.assertIsSatisfied(context);
@@ -40,10 +38,9 @@ public class CordaConsumerVaultTrackByWithSortingTest extends CordaConsumerTestS
             public void configure() {
                 errorHandler(deadLetterChannel("mock:error"));
 
-                from(getUrl() + "&" + OPERATION.toLowerCase() + "=" + VAULT_TRACK_BY_WITH_SORTING
+                from(getUrl() + "&" + OPERATION.toLowerCase() + "=" + VAULT_TRACK_BY_CRITERIA
                         + "&contractStateClass=#contractStateClass"
-                        + "&queryCriteria=#queryCriteria"
-                        + "&sort=#sort")
+                        + "&queryCriteria=#queryCriteria")
                         .to("mock:result");
             }
         };
