+++ components/camel-corda/src/test/java/org/apache/camel/component/corda/CordaConsumerVaultTrackByIntegrationTest.java
@@ -18,17 +18,15 @@ package org.apache.camel.component.corda;
 
 import org.apache.camel.builder.RouteBuilder;
 import org.apache.camel.component.mock.MockEndpoint;
-import org.junit.Ignore;
-import org.junit.Test;
+import org.junit.jupiter.api.Test;
 
 import static org.apache.camel.component.corda.CordaConstants.OPERATION;
-import static org.apache.camel.component.corda.CordaConstants.VAULT_TRACK_BY_CRITERIA;
+import static org.apache.camel.component.corda.CordaConstants.VAULT_TRACK_BY;
 
-@Ignore("This integration test requires a locally running corda node such cordapp-template-java")
-public class CordaConsumerVaultTrackByCriteriaTest extends CordaConsumerTestSupport {
+public class CordaConsumerVaultTrackByIntegrationTest extends CordaConsumerTestSupport {
 
     @Test
-    public void vaultTrackByCriteriaTest() throws Exception {
+    public void vaultTrackByTest() throws Exception {
         mockResult.expectedMinimumMessageCount(1);
         mockError.expectedMessageCount(0);
         MockEndpoint.assertIsSatisfied(context);
@@ -40,9 +38,11 @@ public class CordaConsumerVaultTrackByCriteriaTest extends CordaConsumerTestSupp
             public void configure() {
                 errorHandler(deadLetterChannel("mock:error"));
 
-                from(getUrl() + "&" + OPERATION.toLowerCase() + "=" + VAULT_TRACK_BY_CRITERIA
+                from(getUrl() + "&" + OPERATION.toLowerCase() + "=" + VAULT_TRACK_BY
                         + "&contractStateClass=#contractStateClass"
-                        + "&queryCriteria=#queryCriteria")
+                        + "&queryCriteria=#queryCriteria"
+                        + "&pageSpecification=#pageSpecification"
+                        + "&sort=#sort")
                         .to("mock:result");
             }
         };
