+++ components/camel-corda/src/test/java/org/apache/camel/component/corda/CordaConsumerVaultTrackByWithSortingIntegrationTest.java
@@ -18,17 +18,15 @@ package org.apache.camel.component.corda;
 
 import org.apache.camel.builder.RouteBuilder;
 import org.apache.camel.component.mock.MockEndpoint;
-import org.junit.Ignore;
-import org.junit.Test;
+import org.junit.jupiter.api.Test;
 
 import static org.apache.camel.component.corda.CordaConstants.OPERATION;
-import static org.apache.camel.component.corda.CordaConstants.VAULT_TRACK_BY_WITH_PAGING_SPEC;
+import static org.apache.camel.component.corda.CordaConstants.VAULT_TRACK_BY_WITH_SORTING;
 
-@Ignore("This integration test requires a locally running corda node such cordapp-template-java")
-public class CordaConsumerVaultTrackByWithPagingSpecTest extends CordaConsumerTestSupport {
+public class CordaConsumerVaultTrackByWithSortingIntegrationTest extends CordaConsumerTestSupport {
 
     @Test
-    public void vaultTrackByWithPagingSpecTest() throws Exception {
+    public void vaultTrackByWithSortingTest() throws Exception {
         mockResult.expectedMinimumMessageCount(1);
         mockError.expectedMessageCount(0);
         MockEndpoint.assertIsSatisfied(context);
@@ -40,10 +38,10 @@ public class CordaConsumerVaultTrackByWithPagingSpecTest extends CordaConsumerTe
             public void configure() {
                 errorHandler(deadLetterChannel("mock:error"));
 
-                from(getUrl() + "&" + OPERATION.toLowerCase() + "=" + VAULT_TRACK_BY_WITH_PAGING_SPEC
+                from(getUrl() + "&" + OPERATION.toLowerCase() + "=" + VAULT_TRACK_BY_WITH_SORTING
                         + "&contractStateClass=#contractStateClass"
                         + "&queryCriteria=#queryCriteria"
-                        + "&pageSpecification=#pageSpecification")
+                        + "&sort=#sort")
                         .to("mock:result");
             }
         };
